refactor(pi-sequential-thinking): extract output formatting into output.ts
Move formatToolOutput, writeTempFile, toJsonString, and the McpToolDetails
shape out of index.ts into a focused output.ts module. The functions
remain re-exported from index.ts so existing helper tests keep working
unchanged.

Reduces index.ts by ~100 lines and isolates the truncation/temp-file
logic from tool dispatch.
